2794 - API improvements as per our discussion (#2795)
