+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/avatar/ChannelAvatar.kt
@@ -16,7 +16,6 @@ import io.getstream.chat.android.client.models.initials
 import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.previewdata.PreviewChannelData
 import io.getstream.chat.android.compose.state.OnlineIndicatorAlignment
-import io.getstream.chat.android.compose.ui.components.OnlineIndicator
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 
 /**
@@ -43,7 +42,7 @@ public fun ChannelAvatar(
     showOnlineIndicator: Boolean = true,
     onlineIndicatorAlignment: OnlineIndicatorAlignment = OnlineIndicatorAlignment.TopEnd,
     onlineIndicator: @Composable BoxScope.() -> Unit = {
-        OnlineIndicator(modifier = Modifier.align(onlineIndicatorAlignment.alignment))
+        DefaultOnlineIndicator(onlineIndicatorAlignment)
     },
     contentDescription: String? = null,
     onClick: (() -> Unit)? = null,
