+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/avatar/UserAvatar.kt
@@ -43,7 +43,7 @@ public fun UserAvatar(
     showOnlineIndicator: Boolean = true,
     onlineIndicatorAlignment: OnlineIndicatorAlignment = OnlineIndicatorAlignment.TopEnd,
     onlineIndicator: @Composable BoxScope.() -> Unit = {
-        OnlineIndicator(modifier = Modifier.align(onlineIndicatorAlignment.alignment))
+        DefaultOnlineIndicator(onlineIndicatorAlignment)
     },
     onClick: (() -> Unit)? = null,
 ) {
@@ -84,6 +84,14 @@ public fun UserAvatar(
         // Apply modifier to the avatar itself
         avatarContent(modifier = modifier)
     }
+}
+
+/**
+ * The default online indicator for channel members.
+ */
+@Composable
+internal fun BoxScope.DefaultOnlineIndicator(onlineIndicatorAlignment: OnlineIndicatorAlignment) {
+    OnlineIndicator(modifier = Modifier.align(onlineIndicatorAlignment.alignment))
 }
 
 @Preview
