+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/suggestions/commands/CommandSuggestionItem.kt
@@ -26,7 +26,7 @@ import io.getstream.chat.android.compose.R
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 
 /**
- * Represents the default command suggestion item in the command suggestion list popup.
+ * Represents the command suggestion item in the command suggestion list popup.
  *
  * @param command The command to display.
  * @param modifier Modifier for styling.
@@ -35,24 +35,15 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
  * @param detailsContent Customizable composable function that represents the details content of a command item.
  */
 @Composable
-public fun DefaultCommandSuggestionItem(
+public fun CommandSuggestionItem(
     command: Command,
     modifier: Modifier = Modifier,
     onCommandSelected: (Command) -> Unit = {},
     leadingContent: @Composable RowScope.(Command) -> Unit = {
-        DefaultCommandSuggestionItemLeadingContent(
-            modifier = Modifier
-                .padding(end = 8.dp)
-                .size(ChatTheme.dimens.commandSuggestionItemIconSize)
-        )
+        DefaultCommandSuggestionItemLeadingContent()
     },
     detailsContent: @Composable RowScope.(Command) -> Unit = {
-        DefaultCommandSuggestionItemDetailsContent(
-            command = it,
-            modifier = Modifier
-                .weight(1f)
-                .wrapContentHeight()
-        )
+        DefaultCommandSuggestionItemDetailsContent(command = it)
     },
 ) {
     Row(
@@ -78,15 +69,13 @@ public fun DefaultCommandSuggestionItem(
 
 /**
  * Represents the default content shown at the start of the command list item.
- *
- * @param modifier Modifier for styling.
  */
 @Composable
-public fun DefaultCommandSuggestionItemLeadingContent(
-    modifier: Modifier = Modifier,
-) {
+internal fun DefaultCommandSuggestionItemLeadingContent() {
     Image(
-        modifier = modifier,
+        modifier = Modifier
+            .padding(end = 8.dp)
+            .size(ChatTheme.dimens.commandSuggestionItemIconSize),
         painter = painterResource(id = R.drawable.stream_compose_ic_giphy),
         contentDescription = null
     )
@@ -99,7 +88,7 @@ public fun DefaultCommandSuggestionItemLeadingContent(
  *  @param modifier Modifier for styling.
  */
 @Composable
-public fun DefaultCommandSuggestionItemDetailsContent(
+internal fun RowScope.DefaultCommandSuggestionItemDetailsContent(
     command: Command,
     modifier: Modifier = Modifier,
 ) {
@@ -110,7 +99,9 @@ public fun DefaultCommandSuggestionItemDetailsContent(
     )
 
     Row(
-        modifier = modifier,
+        modifier = modifier
+            .weight(1f)
+            .wrapContentHeight(),
         verticalAlignment = Alignment.CenterVertically
     ) {
         Text(
