+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/suggestions/commands/CommandSuggestionList.kt
@@ -80,3 +80,17 @@ public fun CommandSuggestionList(
         }
     }
 }
+
+/**
+ * The default command suggestion item.
+ *
+ * @param command The given command.
+ * @param onCommandSelected Handler when the command is selected.
+ */
+@Composable
+internal fun DefaultCommandSuggestionItem(
+    command: Command,
+    onCommandSelected: (Command) -> Unit,
+) {
+    CommandSuggestionItem(command = command, onCommandSelected = onCommandSelected)
+}
