+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/suggestions/mentions/MentionSuggestionList.kt
@@ -46,3 +46,17 @@ public fun MentionSuggestionList(
         }
     }
 }
+
+/**
+ * The default mention suggestion item.
+ */
+@Composable
+internal fun DefaultMentionSuggestionItem(
+    user: User,
+    onMentionSelected: (User) -> Unit,
+) {
+    MentionSuggestionItem(
+        user = user,
+        onMentionSelected = onMentionSelected,
+    )
+}
